# backend/api/swagger.yaml
openapi: "3.0.0"
info:
  title: API Music Room Suricates 
  description: This is a MusicRoom server with all the requests available to the users.
    You can authentify yourself to get access to the requests.
  version: 1.0.0
servers:
  - url: "http://62.34.5.191:45559/api"
  - url: "http://localhost:45559/api"
  - url: "http://localhost:46000/api"
...
paths:
...
  /auth/signup/google:
    post:
      tags:
      - "auth"
      summary: "Create an account to MusicRoom with Google"
      operationId: "signupGoogle"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                token:
                  type: string
                MusicRoom_ID:
                  type: string
              required:
                - token
                - MusicRoom_ID
      responses:
        201:
          description: |
            __{code:__ '0'__}__ - `Success - User created - Validation email has been sent!`

            __{code:__ '1'__}__ - `Success - User created - Account is already activated!`
        400:
          description: |
            __{code:__ '0'__}__ - `Email already assign to a MusicRoom account!`
          
            __{code:__ '1'__}__ - `Missing token or MusicRoom_ID field!`

            __{code:__ '2'__}__ - `Auth with Google failed, invalid token or client id!`
        500: 
          description: |
           __{code:__ '0'__}__ - `Server error or Unhandled error!`
      security: []
  /auth/signup/facebook:
    post:
      tags:
      - "auth"
      summary: "Create an account to MusicRoom with Facebook"
      operationId: "signupFacebook"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        201:
          description: |
            __{code:__ '0'__}__ - `Success - User created - Account is already activated!`
        400:
          description: |
            __{code:__ '0'__}__ - `Email already assign to a MusicRoom account!`
          
            __{code:__ '1'__}__ - `Missing token field!`

            __{code:__ '2'__}__ - `Auth with Facebook failed, invalid token!`
        500: 
          description: |
           __{code:__ '0'__}__ - `Server error or Unhandled error!`
      security: []
  /auth/login:
    post:
      tags:
      - "auth"
      summary: "Login to your MusicRoom account"
      operationId: "login"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: |
            __{code:__ '0'__, userId:__ 'userId'__, token:__ 'token'__}__ - `Success user logged!`
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  userId:
                    type: string
                  token:
                    type: string
                example:
                  code: "0"
                  userId: 601014ae2fea2f1bbc9354fc
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3mDEwMTRhZTJmYmEyZjFiYmM5MzU0ZmMiLCJpYXQiOjE2MTE3NTY1MjQsImV4cCI6MTYxMTg0MjkyNH0.GWKNU6Ci0chgUOEsv09pqtmaXytSJKS-l-ZnALD1DXk
        400:
          description: |
            __{code:__ '0'__}__ - `Missing email or password field!`

            __{code:__ '1'__}__ - `Invalid user email!`

            __{code:__ '2'__}__ - `Account is not activated yet!`

            __{code:__ '3'__}__ - `Invalid password!`
        500:
          description: |
            __{code:__ '0'__}__ - `Server error or Unhandled error`
      security: []
  /auth/login/google:
    post:
      tags:
      - "auth"
      summary: "Login to your MusicRoom account with google"
      operationId: "loginGoogle"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                token:
                  type: string
                MusicRoom_ID:
                  type: string
              required:
                - token
                - MusicRoom_ID
      responses:
        200:
          description: |
            __{code:__ '0'__, userId:__ 'userId'__, token:__ 'token'__}__ - `Success user logged!`
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  userId:
                    type: string
                  token:
                    type: string
                example:
                  code: "0"
                  userId: 601014ae2fea2f1bbc9354fc
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3mDEwMTRhZTJmYmEyZjFiYmM5MzU0ZmMiLCJpYXQiOjE2MTE3NTY1MjQsImV4cCI6MTYxMTg0MjkyNH0.GWKNU6Ci0chgUOEsv09pqtmaXytSJKS-l-ZnALD1DXk
        400:
          description: |
            __{code:__ '0'__}__ - `Invalid user email!`

            __{code:__ '1'__}__ - `Account is not activated yet!`

            __{code:__ '2'__}__ - `Missing token or MusicRoom_ID field!`

            __{code:__ '3'__}__ - `Auth with Google failed, invalid token or client id!`
        500:
          description: |
            __{code:__ '0'__}__ - `Server error or Unhandled error`
      security: []
  /auth/login/facebook:
    post:
      tags:
      - "auth"
      summary: "Login to your MusicRoom account with facebook"
      operationId: "loginFacebook"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        200:
          description: |
            __{code:__ '0'__, userId:__ 'userId'__, token:__ 'token'__}__ - `Success user logged!`
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  userId:
                    type: string
                  token:
                    type: string
                example:
                  code: "0"
                  userId: 601014ae2fea2f1bbc9354fc
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI3mDEwMTRhZTJmYmEyZjFiYmM5MzU0ZmMiLCJpYXQiOjE2MTE3NTY1MjQsImV4cCI6MTYxMTg0MjkyNH0.GWKNU6Ci0chgUOEsv09pqtmaXytSJKS-l-ZnALD1DXk
        400:
          description: |
            __{code:__ '0'__}__ - `Invalid user email!`

            __{code:__ '1'__}__ - `Missing token or MusicRoom_ID field!`

            __{code:__ '2'__}__ - `Auth with Google failed, invalid token or client id!`
        500:
          description: |
            __{code:__ '0'__}__ - `Server error or Unhandled error`
      security: []
  /auth/password:
    post:
      tags:
      - "auth"
      summary: "Send a mail to reset a password"
      operationId: "sendMailResetPassword"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        200: 
          description: "Success - Mail send!"
        500: 
          description: "Error - Internal server error. Error object returned."
      security: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []